+++ atlasdb-cassandra/src/test/java/com/palantir/atlasdb/cassandra/backup/TransactionAborterTest.java
@@ -37,11 +37,11 @@ import com.datastax.driver.core.TableMetadata;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Range;
 import com.palantir.atlasdb.AtlasDbConstants;
+import com.palantir.atlasdb.atomic.AtomicValue;
 import com.palantir.atlasdb.cassandra.CassandraKeyValueServiceConfig;
 import com.palantir.atlasdb.cassandra.backup.transaction.TransactionTableEntries;
 import com.palantir.atlasdb.cassandra.backup.transaction.TransactionTableEntry;
 import com.palantir.atlasdb.cassandra.backup.transaction.TransactionsTableInteraction;
-import com.palantir.atlasdb.pue.AtomicValue;
 import com.palantir.atlasdb.timelock.api.Namespace;
 import com.palantir.timestamp.FullyBoundedTimestampRange;
 import java.util.List;
