+++ atlasdb-impl-shared/src/test/java/com/palantir/atlasdb/atomic/ResilientCommitTimestampPutUnlessExistsTableTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.atlasdb.pue;
+package com.palantir.atlasdb.atomic;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
@@ -76,13 +76,13 @@ public class ResilientCommitTimestampPutUnlessExistsTableTest {
                     spiedKvs, TableReference.createFromFullyQualifiedName("test.table")));
 
     private final boolean validating;
-    private final AtomicTable<Long, Long> pueTable;
+    private final AtomicTable<Long, Long> atomicTable;
     private final AtomicLong clockLong = new AtomicLong(1000);
     private final Clock clock = clockLong::get;
 
     public ResilientCommitTimestampPutUnlessExistsTableTest(String name, Object parameter) {
         validating = (boolean) parameter;
-        pueTable = new ResilientCommitTimestampAtomicTable(
+        atomicTable = new ResilientCommitTimestampAtomicTable(
                 spiedStore,
                 TwoPhaseEncodingStrategy.INSTANCE,
                 () -> !validating,
@@ -101,8 +101,8 @@ public class ResilientCommitTimestampPutUnlessExistsTableTest {
 
     @Test
     public void canPutAndGet() throws ExecutionException, InterruptedException {
-        pueTable.update(1L, 2L);
-        assertThat(pueTable.get(1L).get()).isEqualTo(2L);
+        atomicTable.update(1L, 2L);
+        assertThat(atomicTable.get(1L).get()).isEqualTo(2L);
 
         verify(spiedStore).atomicUpdate(anyMap());
         verify(spiedStore, atLeastOnce()).put(anyMap());
@@ -111,29 +111,29 @@ public class ResilientCommitTimestampPutUnlessExistsTableTest {
 
     @Test
     public void emptyReturnsNull() throws ExecutionException, InterruptedException {
-        assertThat(pueTable.get(3L).get()).isNull();
+        assertThat(atomicTable.get(3L).get()).isNull();
     }
 
     @Test
     public void cannotPueTwice() {
-        pueTable.update(1L, 2L);
-        assertThatThrownBy(() -> pueTable.update(1L, 2L)).isInstanceOf(KeyAlreadyExistsException.class);
+        atomicTable.update(1L, 2L);
+        assertThatThrownBy(() -> atomicTable.update(1L, 2L)).isInstanceOf(KeyAlreadyExistsException.class);
     }
 
     @Test
     public void canPutAndGetMultiple() throws ExecutionException, InterruptedException {
         ImmutableMap<Long, Long> inputs = ImmutableMap.of(1L, 2L, 3L, 4L, 7L, 8L);
-        pueTable.updateMultiple(inputs);
-        assertThat(pueTable.get(ImmutableList.of(1L, 3L, 5L, 7L)).get()).containsExactlyInAnyOrderEntriesOf(inputs);
+        atomicTable.updateMultiple(inputs);
+        assertThat(atomicTable.get(ImmutableList.of(1L, 3L, 5L, 7L)).get()).containsExactlyInAnyOrderEntriesOf(inputs);
     }
 
     @Test
     public void pueThatThrowsIsCorrectedOnGet() throws ExecutionException, InterruptedException {
         spiedStore.startFailingPuts();
-        assertThatThrownBy(() -> pueTable.update(1L, 2L)).isInstanceOf(RuntimeException.class);
+        assertThatThrownBy(() -> atomicTable.update(1L, 2L)).isInstanceOf(RuntimeException.class);
         spiedStore.stopFailingPuts();
 
-        assertThat(pueTable.get(1L).get()).isEqualTo(2L);
+        assertThat(atomicTable.get(1L).get()).isEqualTo(2L);
         verify(spiedStore, times(2)).put(anyMap());
     }
 
@@ -157,7 +157,7 @@ public class ResilientCommitTimestampPutUnlessExistsTableTest {
                 .when(spiedStore)
                 .checkAndTouch(timestampAsCell, stagingValue);
 
-        assertThat(pueTable.get(startTimestamp).get()).isEqualTo(commitTimestamp);
+        assertThat(atomicTable.get(startTimestamp).get()).isEqualTo(commitTimestamp);
     }
 
     @Test
@@ -193,7 +193,7 @@ public class ResilientCommitTimestampPutUnlessExistsTableTest {
 
         int numberOfReads = 100;
         for (int i = 0; i < numberOfReads; i++) {
-            assertThatThrownBy(() -> pueTable.get(0L).get())
+            assertThatThrownBy(() -> atomicTable.get(0L).get())
                     .isInstanceOf(ExecutionException.class)
                     .hasCauseInstanceOf(RuntimeException.class)
                     .hasMessageContaining("Failed to set value");
@@ -208,7 +208,7 @@ public class ResilientCommitTimestampPutUnlessExistsTableTest {
 
         spiedStore.stopFailingPuts();
         for (long i = 0; i < 100; i++) {
-            assertThat(pueTable.get(0L).get()).isEqualTo(0L);
+            assertThat(atomicTable.get(0L).get()).isEqualTo(0L);
         }
 
         if (validating) {
@@ -229,7 +229,7 @@ public class ResilientCommitTimestampPutUnlessExistsTableTest {
         List<Future<ListenableFuture<Long>>> results = new ArrayList<>();
         int numberOfReads = 20_000;
         for (int i = 0; i < numberOfReads; i++) {
-            results.add(writers.submit(() -> pueTable.get(random.nextLong(50L))));
+            results.add(writers.submit(() -> atomicTable.get(random.nextLong(50L))));
         }
         results.forEach(future -> assertThatCode(() -> future.get().get()).doesNotThrowAnyException());
         writers.shutdownNow();
@@ -254,7 +254,7 @@ public class ResilientCommitTimestampPutUnlessExistsTableTest {
         int numberOfReads = parallelism * 100;
         for (int i = 0; i < numberOfReads; i++) {
             // all timestamps go into the same row
-            results.add(Futures.submitAsync(() -> pueTable.get(random.nextLong(10) * rowsPerQuantum), writers));
+            results.add(Futures.submitAsync(() -> atomicTable.get(random.nextLong(10) * rowsPerQuantum), writers));
         }
         results.forEach(future -> assertThatThrownBy(future::get)
                 .isInstanceOf(ExecutionException.class)
@@ -262,7 +262,7 @@ public class ResilientCommitTimestampPutUnlessExistsTableTest {
         spiedStore.stopFailingPuts();
         results.clear();
         for (int i = 0; i < numberOfReads; i++) {
-            results.add(Futures.submitAsync(() -> pueTable.get(random.nextLong(10) * rowsPerQuantum), writers));
+            results.add(Futures.submitAsync(() -> atomicTable.get(random.nextLong(10) * rowsPerQuantum), writers));
         }
         results.forEach(future -> assertThatCode(future::get).doesNotThrowAnyException());
         writers.shutdownNow();
@@ -280,7 +280,7 @@ public class ResilientCommitTimestampPutUnlessExistsTableTest {
         List<ListenableFuture<Long>> results = new ArrayList<>();
         int numberOfReads = maximumParallelism * 100;
         for (int i = 0; i < numberOfReads; i++) {
-            results.add(Futures.submitAsync(() -> pueTable.get(random.nextLong(maximumParallelism * 2)), writers));
+            results.add(Futures.submitAsync(() -> atomicTable.get(random.nextLong(maximumParallelism * 2)), writers));
         }
         results.forEach(future -> assertThatThrownBy(future::get)
                 .isInstanceOf(ExecutionException.class)
@@ -288,7 +288,7 @@ public class ResilientCommitTimestampPutUnlessExistsTableTest {
         spiedStore.stopFailingPuts();
         results.clear();
         for (int i = 0; i < numberOfReads; i++) {
-            results.add(Futures.submitAsync(() -> pueTable.get(random.nextLong(maximumParallelism * 2)), writers));
+            results.add(Futures.submitAsync(() -> atomicTable.get(random.nextLong(maximumParallelism * 2)), writers));
         }
         results.forEach(future -> assertThatCode(future::get).doesNotThrowAnyException());
         writers.shutdownNow();
@@ -307,7 +307,7 @@ public class ResilientCommitTimestampPutUnlessExistsTableTest {
         setupStagingValues(1);
         spiedStore.enableCommittingUnderUs();
 
-        assertThat(pueTable.get(0L).get()).isEqualTo(0L);
+        assertThat(atomicTable.get(0L).get()).isEqualTo(0L);
         verify(spiedKvs, times(1)).checkAndSet(any());
         // only the put from the original PUE was registered
         verify(spiedStore, times(1)).put(anyMap());
@@ -320,15 +320,15 @@ public class ResilientCommitTimestampPutUnlessExistsTableTest {
         spiedStore.stopFailingPuts();
         spiedStore.startSlowPue();
 
-        assertThat(pueTable.get(0L).get()).isEqualTo(0L);
-        assertThat(pueTable.get(1L).get()).isEqualTo(1L);
+        assertThat(atomicTable.get(0L).get()).isEqualTo(0L);
+        assertThat(atomicTable.get(1L).get()).isEqualTo(1L);
 
         verify(spiedKvs, times(1)).checkAndSet(any());
         verify(spiedStore, times(1 + 2)).put(anyMap());
 
         clockLong.accumulateAndGet(Duration.ofSeconds(62).toMillis(), Long::sum);
 
-        assertThat(pueTable.get(2L).get()).isEqualTo(2L);
+        assertThat(atomicTable.get(2L).get()).isEqualTo(2L);
         verify(spiedKvs, times(2)).checkAndSet(any());
         verify(spiedStore, times(1 + 3)).put(anyMap());
     }
@@ -339,18 +339,18 @@ public class ResilientCommitTimestampPutUnlessExistsTableTest {
         setupStagingValues(5);
         spiedStore.failPutsWithAtlasdbDependencyException();
 
-        assertThatThrownBy(() -> pueTable.get(0L).get())
+        assertThatThrownBy(() -> atomicTable.get(0L).get())
                 .isInstanceOf(ExecutionException.class)
                 .hasCauseInstanceOf(RetryLimitReachedException.class);
         spiedStore.stopFailingPuts();
 
-        assertThat(pueTable.get(1L).get()).isEqualTo(1L);
+        assertThat(atomicTable.get(1L).get()).isEqualTo(1L);
         verify(spiedKvs, times(1)).checkAndSet(any());
         verify(spiedStore, times(1 + 2)).put(anyMap());
 
         clockLong.accumulateAndGet(Duration.ofSeconds(62).toMillis(), Long::sum);
 
-        assertThat(pueTable.get(2L).get()).isEqualTo(2L);
+        assertThat(atomicTable.get(2L).get()).isEqualTo(2L);
         verify(spiedKvs, times(2)).checkAndSet(any());
         verify(spiedStore, times(1 + 3)).put(anyMap());
     }
@@ -358,7 +358,7 @@ public class ResilientCommitTimestampPutUnlessExistsTableTest {
     private void setupStagingValues(int num) {
         spiedStore.startFailingPuts();
         Map<Long, Long> initialWrites = LongStream.range(0, num).boxed().collect(Collectors.toMap(x -> x, x -> x));
-        assertThatThrownBy(() -> pueTable.updateMultiple(initialWrites))
+        assertThatThrownBy(() -> atomicTable.updateMultiple(initialWrites))
                 .isInstanceOf(RuntimeException.class)
                 .hasMessageContaining("Failed to set value");
     }
